python3/requirements.txt
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
flake8==6.1.0
idna==3.4
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
packaging==23.2
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
pyotp==2.9.0
pyrh==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.31.0
robin-stocks==3.0.6
six==1.16.0
typing==3.7.4.3
urllib3==2.0.7
yarl==1.9.2
